+++ dataframe-jdbc/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/db/H2.kt
@@ -1,17 +1,22 @@
 package org.jetbrains.kotlinx.dataframe.io.db
 
-import org.jetbrains.kotlinx.dataframe.io.JdbcColumn
+import org.jetbrains.kotlinx.dataframe.io.TableColumnMetadata
 import org.jetbrains.kotlinx.dataframe.schema.ColumnSchema
 import java.sql.ResultSet
 import kotlin.reflect.typeOf
 
 /**
- * NOTE: all date and timestamp related types converted to String to avoid java.sql.* types
+ * Represents the H2 database type.
+ *
+ * This class provides methods to convert data from a ResultSet to the appropriate type for H2,
+ * and to generate the corresponding column schema.
+ *
+ * NOTE: All date and timestamp related types are converted to String to avoid java.sql.* types.
  */
 public object H2 : DbType("h2") {
-    override fun convertDataFromResultSet(rs: ResultSet, jdbcColumn: JdbcColumn): Any? {
-        val name = jdbcColumn.name
-        return when (jdbcColumn.sqlType) {
+    override fun convertDataFromResultSet(rs: ResultSet, tableColumn: TableColumnMetadata): Any? {
+        val name = tableColumn.name
+        return when (tableColumn.sqlType) {
             "CHARACTER", "CHAR" -> rs.getString(name)
             "CHARACTER VARYING", "CHAR VARYING",  "VARCHAR" -> rs.getString(name)
             "CHARACTER LARGE OBJECT", "CHAR LARGE OBJECT", "CLOB" -> rs.getString(name)
@@ -40,12 +45,12 @@ public object H2 : DbType("h2") {
             "JSON" -> rs.getString(name)
             "UUID" -> rs.getString(name)
             //"ARRAY" -> rs.getArray(jdbcColumn.name)
-            else -> throw IllegalArgumentException("Unsupported H2 type: ${jdbcColumn.sqlType}")
+            else -> throw IllegalArgumentException("Unsupported H2 type: ${tableColumn.sqlType}")
         }
     }
 
-    override fun toColumnSchema(jdbcColumn: JdbcColumn): ColumnSchema {
-        return when (jdbcColumn.sqlType) {
+    override fun toColumnSchema(tableColumn: TableColumnMetadata): ColumnSchema {
+        return when (tableColumn.sqlType) {
             "CHARACTER", "CHAR" -> ColumnSchema.Value(typeOf<String>())
             "CHARACTER VARYING", "CHAR VARYING",  "VARCHAR" -> ColumnSchema.Value(typeOf<String>())
             "CHARACTER LARGE OBJECT", "CHAR LARGE OBJECT", "CLOB" -> ColumnSchema.Value(typeOf<String>())
@@ -74,7 +79,7 @@ public object H2 : DbType("h2") {
             "JSON" -> ColumnSchema.Value(typeOf<String>())
             "UUID" -> ColumnSchema.Value(typeOf<String>())
             //"ARRAY" -> rs.getArray(jdbcColumn.name)
-            else -> throw IllegalArgumentException("Unsupported H2 type: ${jdbcColumn.sqlType}")
+            else -> throw IllegalArgumentException("Unsupported H2 type: ${tableColumn.sqlType}")
         }
     }
 }
