+++ dataframe-jdbc/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/db/MariaDb.kt
@@ -1,14 +1,20 @@
 package org.jetbrains.kotlinx.dataframe.io.db
 
-import org.jetbrains.kotlinx.dataframe.io.JdbcColumn
+import org.jetbrains.kotlinx.dataframe.io.TableColumnMetadata
 import org.jetbrains.kotlinx.dataframe.schema.ColumnSchema
 import java.sql.ResultSet
 import kotlin.reflect.typeOf
 
+/**
+ * Represents the MariaDb database type.
+ *
+ * This class provides methods to convert data from a ResultSet to the appropriate type for MariaDb,
+ * and to generate the corresponding column schema.
+ */
 public object MariaDb : DbType("mariadb") {
-    override fun convertDataFromResultSet(rs: ResultSet, jdbcColumn: JdbcColumn): Any? {
-        val name = jdbcColumn.name
-        return when (jdbcColumn.sqlType) {
+    override fun convertDataFromResultSet(rs: ResultSet, tableColumn: TableColumnMetadata): Any? {
+        val name = tableColumn.name
+        return when (tableColumn.sqlType) {
             "BIT" -> rs.getBytes(name)
             "TINYINT" -> rs.getInt(name)
             "SMALLINT" -> rs.getInt(name)
@@ -37,12 +43,12 @@ public object MariaDb : DbType("mariadb") {
             "LONGTEXT" -> rs.getString(name)
             "ENUM" -> rs.getString(name)
             "SET" -> rs.getString(name)
-            else -> throw IllegalArgumentException("Unsupported MariaDB type: ${jdbcColumn.sqlType}")
+            else -> throw IllegalArgumentException("Unsupported MariaDB type: ${tableColumn.sqlType}")
         }
     }
 
-    override fun toColumnSchema(jdbcColumn: JdbcColumn): ColumnSchema {
-        return when (jdbcColumn.sqlType) {
+    override fun toColumnSchema(tableColumn: TableColumnMetadata): ColumnSchema {
+        return when (tableColumn.sqlType) {
             "BIT" -> ColumnSchema.Value(typeOf<ByteArray>())
             "TINYINT" -> ColumnSchema.Value(typeOf<Int>())
             "SMALLINT" -> ColumnSchema.Value(typeOf<Int>())
@@ -71,7 +77,7 @@ public object MariaDb : DbType("mariadb") {
             "LONGTEXT" -> ColumnSchema.Value(typeOf<String>())
             "ENUM" -> ColumnSchema.Value(typeOf<String>())
             "SET" -> ColumnSchema.Value(typeOf<String>())
-            else -> throw IllegalArgumentException("Unsupported MariaDB type: ${jdbcColumn.sqlType}")
+            else -> throw IllegalArgumentException("Unsupported MariaDB type: ${tableColumn.sqlType}")
         }
     }
 }
