+++ dataframe-jdbc/src/main/resources/META-INF/services/org.jetbrains.kotlinx.dataframe.io.SupportedFormat
@@ -1,1 +1,1 @@
-org.jetbrains.kotlinx.dataframe.io.JDBC
+org.jetbrains.kotlinx.dataframe.io.Jdbc
