+++ dataframe-jdbc/src/test/kotlin/org/jetbrains/kotlinx/dataframe/io/h2Test.kt
@@ -70,7 +70,7 @@ interface TestTableData {
     val uuidCol: String
 }
 
-class JDBCTest {
+class JdbcTest {
     companion object {
         private lateinit var connection: Connection
